fix death count
